[Modify] GlobalStyles 버튼 배경 투명화 추가
